requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-celery-results==1.2.1
django-filter==2.4.0
django-nested-inline==0.4.2
django-phonenumber-field==5.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.12.12
postgres==3.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyparsing==2.4.7
pytz==2020.4
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.2.5
